Implement stream cancellation in HTTP/3 QPACK

+++ jetty-http3/http3-qpack/src/main/java/org/eclipse/jetty/http3/qpack/internal/instruction/StreamCancellationInstruction.java
@@ -22,9 +22,9 @@ import org.eclipse.jetty.util.BufferUtil;
 
 public class StreamCancellationInstruction implements Instruction
 {
-    private final int _streamId;
+    private final long _streamId;
 
-    public StreamCancellationInstruction(int streamId)
+    public StreamCancellationInstruction(long streamId)
     {
         _streamId = streamId;
     }
